pyproject.toml
# NOEMA • pyproject.toml — پیکربندی مینیمال بسته و بیلد

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "noema"
version = "0.1.0"
description = "NOEMA — اسکلت مغز نوما: I/O متنی، مهارت‌ها، رجیستری ابزار، ایمنی، خواب/تثبیت."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "NOMA/NOEMA Project", email = "noema@example.com" }]
keywords = ["noema", "noma", "cognitive-architecture", "rl", "world-model", "agent"]

# وابستگی‌های هسته (عمداً سبک نگه داشته شده)
dependencies = [
  "PyYAML>=6.0" # اختیاری است، ولی برای خواندن YAMLها مفید است
]

# اکستراها برای قابلیت‌های افزوده (نصب اختیاری)
[project.optional-dependencies]
faiss = ["faiss-cpu>=1.7.4"]
concept = ["scikit-learn>=1.4", "numpy>=1.24"]
data = ["pandas>=2.0", "pyarrow>=15.0"]
dev = ["black>=24.3", "isort>=5.13", "pytest>=8.0"]

[project.urls]
Homepage = "https://example.com/noema"
Repository = "https://example.com/noema/repo"

# اسکریپت‌های خط فرمان سبک (اختیاری)
[project.scripts]
noema-tests = "tests.test_runner:main"
noema-sleep = "sleep.init:run_once"

# -------------------- Setuptools --------------------

[tool.setuptools]
zip-safe = false
include-package-data = true

# الگوی یافتن پکیج‌ها (پوشه‌های کد پروژه)
[tool.setuptools.packages.find]
include = [
  "control*",
  "toolhub*",
  "lang*",
  "selfmeta*",
  "sleep*",
  "safety*",
  "env*",
  "skills*",
  # پوشه‌های رزروی/آتی:
  # "memory*", "concept*", "world*"
]
exclude = ["data*", "scripts*", "tests*"]

# فایل‌های غیرباینری که باید داخل بسته بیایند
[tool.setuptools.package-data]
skills = ["manifest.yaml", "*.py"]
config = ["*.yaml"]
tests = ["regression/*.yaml"]

# -------------------- ابزارهای توسعه (اختیاری) --------------------

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
